requirements.txt
pymilvus
python-dotenv
openai
PyPDF2
python-docx
fastapi
uvicorn
streamlit
requests
